Fix data structure
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@117 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Frameworks/LuceneKit/Source/Index/LCFilterIndexReader.m
@@ -90,7 +90,7 @@
   return [input term];
 }
 
-- (int) docFreq
+- (long) docFreq
 {
   return [input docFreq];
 }
@@ -186,7 +186,7 @@
   return [input termsWithTerm: t];
 }
 
-- (int) docFreq: (LCTerm *) t
+- (long) docFreq: (LCTerm *) t
 {
   return [input docFreq: t];
 }
